Updated Xeogin's Autoexec to 2021-10-10
Copy and pasting his explanation of the changes:

It's only a handful of changes, such as dropping the interp more, but bumping up the smoothtime to make up for it, setting the cmdrate to 100 since that's the new max, got rid of the r_flashlightdepthtexture as it was supposed to be r_flashlightdepthres anyhow and I wanna do more research on those settings anyway, cl_fasttempentcollision got bumped to a crazy value to effectively disable it, dropped snd_mixahead to 0.5 as it should be safe, also bumped fps_max to 400 since monitor refreshrates are getting higher again. Also removed r_pixelvisibility_partial as it's no longer used in mastercom's comfig and I don't know why yet.

80% of the changes are related to mastercom's comfig updates

In a way I'm shifting from super safe for everyone to more optimized